Fix fold-left reduction vectorization with multiple stmt copies
There's a typo when code generating the mask operand for conditional
fold-left reductions in the case we have multiple stmt copies.  The
latter is now allowed for SLP and possibly disabled for non-SLP by
accident.

This fixes the observed run-FAIL for
gcc.dg/vect/vect-cond-reduc-in-order-2-signed-zero.c with AVX512
and 256bit sized vectors.

	* tree-vect-loop.cc (vectorize_fold_left_reduction): Fix
	mask vector operand indexing.