[committed] [RISC-V] Fix false-positive uninitialized variable
Andreas noted we were getting an uninit warning after the recent constant
synthesis changes.  Essentially there's no way for the uninit analysis code to
know the first entry in the CODES array is a UNKNOWN which will set X before
its first use.

So trivial initialization with NULL_RTX is the obvious fix.

Pushed to the trunk.

gcc/

	* config/riscv/riscv.cc (riscv_move_integer): Initialize "x".