[committed] [RISC-V] Drop dead round_32 test
This test is no longer useful.  It doesn't test what it was originally intended
to test and there's really no way to recover it sanely.

We agreed in the patchwork meeting last week that if we want to test Zfa that
we'll write a new test for that.  Similarly if we want to do deeper testing of
the non-Zfa sequences in this space that we'd write new tests for those as well
(execution tests in particular).

So dropping this test.

gcc/testsuite
	* gcc.target/riscv/round_32.c: Delete.