[testsuite] [arm] test board cflags in multilib.exp
multilib.exp tests for multilib-altering flags in a board's
multilib_flags and skips the test, but if such flags appear in the
board's cflags, with the same distorting effects on tested multilibs,
we fail to skip the test.

Extend the skipping logic to board's cflags as well.


for  gcc/testsuite/ChangeLog

	* gcc.target/arm/multilib.exp: Skip based on board cflags too.