libgomp: change alloc-pinned tests failure mode
The feature doesn't work on non-Linux hosts, at present, so skip the tests
entirely.

On Linux systems that have insufficient lockable memory configured we still
need to fail or else the feature won't be getting tested when we think it is,
but now there's a message to explain why.

libgomp/ChangeLog:

	* testsuite/libgomp.c/alloc-pinned-1.c: Change dg-xfail-run-if to
	dg-skip-if.
	Correct spelling mistake.
	Abort on insufficient lockable memory.
	Use #error on non-linux hosts.
	* testsuite/libgomp.c/alloc-pinned-2.c: Likewise.