c++/modules: multiple usings of the same decl [PR115194]
add_binding_entity creates an OVERLOAD to represent a using-declaration in
module purview of a declaration in the global module, even for
non-functions, and we were failing to merge that with the original
declaration in name lookup.

It's not clear to me that building the OVERLOAD is what should be happening,
but let's work around it for now pending an overhaul of using-decl handling
for c++/114683.

	PR c++/115194

gcc/cp/ChangeLog:

	* name-lookup.cc (name_lookup::process_module_binding): Strip an
	OVERLOAD from a non-function.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/using-23_a.C: New test.
	* g++.dg/modules/using-23_b.C: New test.