expand: constify sepops operand to expand_expr_real_2 and expand_widen_pattern_expr [PR113212]
While working on an expand patch back in January I noticed that
the first argument (of sepops type) of expand_expr_real_2 could be
constified as it was not to be touched by the function (nor should it be).
There is code in internal-fn.cc that depends on expand_expr_real_2 not touching
the ops argument so constification makes this more obvious.

Bootstrapped and tested on x86_64-linux-gnu with no regressions.

gcc/ChangeLog:

	PR middle-end/113212
	* expr.h (const_seqpops): New typedef.
	(expand_expr_real_2): Constify the first argument.
	* optabs.cc (expand_widen_pattern_expr): Likewise.
	* optabs.h (expand_widen_pattern_expr): Likewise.
	* expr.cc (expand_expr_real_2):  Likewise
	(do_store_flag): Likewise. Remove incorrect store to ops->code.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>